move qos into v1 module
